refactor: move screens type definitions to screens folder
