+++ src/App.tsx
@@ -6,7 +6,7 @@ import withContexts from './components/with-contexts'
 import { searchScreen } from './screens/Search'
 import { gurbaniScreen } from './screens/Gurbani'
 import { bookmarksScreen } from './screens/Bookmarks'
-import { AppStackParams } from './lib/screens'
+import { AppStackParams } from './screens/screens'
 
 const screens = [ gurbaniScreen, searchScreen, bookmarksScreen ]
