+++ src/screens/Gurbani/index.tsx
@@ -2,7 +2,7 @@ import React from 'react'
 import { StackScreenProps } from '@react-navigation/stack'
 import { useQuery } from 'react-query'
 
-import Screens, { AppStackParams } from '../../lib/screens'
+import Screens, { AppStackParams } from '../screens'
 import { ContentType, LineData } from '../../types/data'
 import { getBookmark, getShabad } from '../../data'
 import Container from '../../components/Container'
