+++ src/screens/screens.ts
@@ -1,0 +1,17 @@
+import { ContentType } from '../types/data'
+import { Override } from '../types/utils'
+
+enum Screens {
+  Home = 'Home',
+  Search = 'Search',
+  Bookmarks = 'Bookmarks',
+  Gurbani = 'Gurbani',
+}
+
+export type AppStackParams = Override<{ [screen in Screens]: undefined }, {
+  [Screens.Gurbani]: { id: string, type: ContentType },
+  [Screens.Bookmarks]: { path: string },
+  [Screens.Search]: undefined,
+}>
+
+export default Screens
